Randomize movement of unhappy cells.
Originally, we would move cells from top to bottom of the grid. The
result was that after the first movement step of unhappy cells, the
bottom of the grid became primarily empty. As cells clumped together,
they tended to do so at the top of the grid. Only during simulations
where there were an approximately even number of both cell types would
we end up with a roughly even distribution.

This change selects unhappy cells at random during the move step
thereby ensuring that the empty cells continue to be randomly
distributed.